requirements.txt
google-generativeai
streamlit
fpdf2
